move search_filter to its own mod
