+++ core/src/blocks/data_source.rs
@@ -1,7 +1,8 @@
 use crate::blocks::block::{
     parse_pair, replace_variables_in_string, Block, BlockResult, BlockType, Env,
 };
-use crate::data_sources::data_source::{Document, SearchFilter};
+use crate::data_sources::data_source::Document;
+use crate::documents_search_filter::SearchFilter;
 use crate::project::Project;
 use crate::Rule;
 use anyhow::{anyhow, Result};
