+++ core/src/lib.rs
@@ -56,3 +56,4 @@ pub mod deno {
 }
 
 pub mod consts;
+pub mod documents_search_filter;
