+++ core/src/stores/store.rs
@@ -93,6 +93,9 @@ pub trait Store {
         document_id: &str,
         version_hash: &Option<String>,
     ) -> Result<Option<Document>>;
+    // async fn find_data_source_documents(
+
+    // )
     async fn upsert_data_source_document(
         &self,
         project: &Project,
